Add build user
The build is running things as root! There doesn't seem to be a way to change the default user.

This adds a build user and some debug lines in the pre_script to show the current user (before switching). Each task has to be run under su. This can be fixed later by using a wrapper script.